requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
sqlalchemy==2.0.32
aiosqlite==0.20.0
jinja2==3.1.4
gpiozero==2.0.1
pyyaml==6.0.2
